+++ heylogs-api/src/main/java/nbbrd/heylogs/Failure.java
@@ -1,5 +1,7 @@
 package nbbrd.heylogs;
 
+import java.util.Objects;
+
 import com.vladsch.flexmark.util.ast.Node;
 import lombok.NonNull;
 import nbbrd.heylogs.spi.Rule;
@@ -28,5 +30,20 @@ public class Failure {
             return line(location.getStartLineNumber() + 1)
                     .column(location.lineColumnAtStart().getSecond() + 1);
         }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
+        Failure failure = (Failure) o;
+        return line == failure.line && column == failure.column && Objects.equals(ruleId, failure.ruleId) && Objects.equals(message, failure.message);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(ruleId, message, line, column);
     }
 }